Add license field to @gluestack-ui/* scoped packages
The license field wasnt being published in the newer @gluestack-ui/*
scoped packages. I assume the license to be applied to these packages
would be the same as the code in the rest of the repo, which is
currently set to MIT license. I copied this attribute to all the
packages manually, and bumped the version.